refactor: update directory name and import
